+++ stroom-integrationtest/src/integrationTest/java/stroom/streamtask/server/TestProxyAggregationTask.java
@@ -103,8 +103,9 @@ public class TestProxyAggregationTask extends AbstractCoreIntegrationTest {
                 filePackProcessorProvider,
                 proxyDir,
                 10,
-                maxAggregation,
                 10000,
+                10000,
+                maxAggregation,
                 maxStreamSize);
         proxyAggregationExecutor.exec(new DummyTask());
     }